YUK-15803 C++ optimizations Removed unused PointData.exemptionStatus and .limit fields Removed TAG_POINT_MAY_BE_EXEMPTED, changed to TAG_POINT_RESERVED Shrunk PointData.type and .quality to byte Shrunk PointData.millis to i16

+++ yukon-client/common/src/com/cannontech/messaging/serialization/thrift/serializer/MessageSerializer.java
@@ -44,7 +44,7 @@ public class MessageSerializer extends
     protected void
         populateThriftEntityFromMessage(ThriftMessageFactory msgFactory, Message msg,
                                         com.cannontech.messaging.serialization.thrift.generated.Message entity) {
-        entity.set_messagePriority(msg.getPriority());
+        entity.set_messagePriority((byte)msg.getPriority());
         entity.set_messageTime(ConverterHelper.dateToMillisec(msg.getTimeStamp()));
         entity.set_soe(msg.getSOE_Tag());
         entity.set_src(msg.getSource());
